+++ Runtime/Timeline/ICommandTimeline.cs
@@ -2,12 +2,12 @@
 {
     public interface ICommandTimeline
     {
-        void ExecuteCommands(long tick);
+        void ExecuteCommands(int tick);
     }
 
     public interface ICommandTimeline<TCommand> : ICommandTimeline
     {
-        void RemoveCommand(long tick, EntityId entityId);
-        void InsertCommand(long tick, in TCommand command, EntityId entityId);
+        void RemoveCommand(int tick, EntityId entityId);
+        void InsertCommand(int tick, in TCommand command, EntityId entityId);
     }
 }
